+++ src/AnomalyDetectors/IsolationForest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\AnomalyDetectors;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Parallel;
 use Rubix\ML\DataType;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
@@ -12,6 +13,9 @@ use Rubix\ML\Helpers\Params;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Graph\Nodes\Depth;
 use Rubix\ML\Graph\Trees\ITree;
+use Rubix\ML\Backends\Serial;
+use Rubix\ML\Backends\Tasks\Task;
+use Rubix\ML\Traits\Multiprocessing;
 use Rubix\ML\Traits\AutotrackRevisions;
 use Rubix\ML\Specifications\DatasetIsNotEmpty;
 use Rubix\ML\Specifications\SpecificationChain;
@@ -20,7 +24,8 @@ use Rubix\ML\Specifications\SamplesAreCompatibleWithEstimator;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
 
-use function count;
+use function array_merge;
+use function ceil;
 
 use const Rubix\ML\EPSILON;
 
@@ -41,9 +46,9 @@ use const Rubix\ML\EPSILON;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class IsolationForest implements Estimator, Learner, Scoring, Persistable
+class IsolationForest implements Estimator, Learner, Scoring, Parallel, Persistable
 {
-    use AutotrackRevisions;
+    use AutotrackRevisions, Multiprocessing;
 
     /**
      * The default minimum anomaly score for a sample to be flagged.
@@ -118,6 +123,52 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
     protected ?int $featureCount = null;
 
     /**
+     * Grow an isolation tree on a unique subset of the training set.
+     *
+     * @internal
+     *
+     * @param Dataset $subset
+     * @param int $maxHeight
+     * @return ITree
+     */
+    public static function growTree(Dataset $subset, int $maxHeight) : ITree
+    {
+        $tree = new ITree($maxHeight);
+
+        $tree->grow($subset);
+
+        return $tree;
+    }
+
+    /**
+     * Predict a chunk of samples.
+     *
+     * @internal
+     *
+     * @param self $estimator
+     * @param Dataset $dataset
+     * @return list<int>
+     */
+    public static function predictChunk(self $estimator, Dataset $dataset) : array
+    {
+        return array_map([$estimator, 'predictSample'], $dataset->samples());
+    }
+
+    /**
+     * Score a chunk of samples.
+     *
+     * @internal
+     *
+     * @param self $estimator
+     * @param Dataset $dataset
+     * @return list<float>
+     */
+    public static function scoreChunk(self $estimator, Dataset $dataset) : array
+    {
+        return array_map([$estimator, 'isolationScore'], $dataset->samples());
+    }
+
+    /**
      * @param int $estimators
      * @param float|null $ratio
      * @param float|null $contamination
@@ -143,6 +194,7 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
         $this->estimators = $estimators;
         $this->ratio = $ratio;
         $this->contamination = $contamination;
+        $this->backend = new Serial();
     }
 
     /**
@@ -218,22 +270,20 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
 
         $maxHeight = (int) max(1, round(log($p, 2.0)));
 
-        $this->trees = [];
+        $this->backend->flush();
 
-        while (count($this->trees) < $this->estimators) {
-            $tree = new ITree($maxHeight);
-
+        for ($i = 0; $i < $this->estimators; ++$i) {
             $subset = $dataset->randomSubset($p);
 
-            $tree->grow($subset);
-
-            $this->trees[] = $tree;
+            $this->backend->enqueue(new Task([self::class, 'growTree'], [$subset, $maxHeight]));
         }
+
+        $this->trees = $this->backend->process();
 
         $this->delta = $this->estimators * Depth::c($p);
 
         if (isset($this->contamination)) {
-            $scores = array_map([$this, 'isolationScore'], $dataset->samples());
+            $scores = $this->scores($dataset);
 
             $threshold = Stats::quantile($scores, 1.0 - $this->contamination);
         }
@@ -258,7 +308,7 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
 
         DatasetHasDimensionality::with($dataset, $this->featureCount)->check();
 
-        return array_map([$this, 'predictSample'], $dataset->samples());
+        return $this->predictions($dataset);
     }
 
     /**
@@ -289,7 +339,7 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
 
         DatasetHasDimensionality::with($dataset, $this->featureCount)->check();
 
-        return array_map([$this, 'isolationScore'], $dataset->samples());
+        return $this->scores($dataset);
     }
 
     /**
@@ -311,6 +361,58 @@ class IsolationForest implements Estimator, Learner, Scoring, Persistable
         $depth /= $this->delta;
 
         return 2.0 ** -$depth;
+    }
+
+    /**
+     * Make predictions on the samples in a dataset in chunks.
+     *
+     * @param Dataset $dataset
+     * @return list<int>
+     */
+    protected function predictions(Dataset $dataset) : array
+    {
+        $chunkSize = (int) ceil($dataset->numSamples() / $this->backend->workers());
+
+        $this->backend->flush();
+
+        foreach ($dataset->batch($chunkSize) as $chunk) {
+            $this->backend->enqueue(new Task([self::class, 'predictChunk'], [$this, $chunk]));
+        }
+
+        $predictions = [];
+
+        foreach ($this->backend->process() as $output) {
+            /** @var list<int> $output */
+            $predictions = array_merge($predictions, $output);
+        }
+
+        return $predictions;
+    }
+
+    /**
+     * Compute the isolation scores of the samples in a dataset in chunks.
+     *
+     * @param Dataset $dataset
+     * @return list<float>
+     */
+    protected function scores(Dataset $dataset) : array
+    {
+        $chunkSize = (int) ceil($dataset->numSamples() / $this->backend->workers());
+
+        $this->backend->flush();
+
+        foreach ($dataset->batch($chunkSize) as $chunk) {
+            $this->backend->enqueue(new Task([self::class, 'scoreChunk'], [$this, $chunk]));
+        }
+
+        $scores = [];
+
+        foreach ($this->backend->process() as $output) {
+            /** @var list<float> $output */
+            $scores = array_merge($scores, $output);
+        }
+
+        return $scores;
     }
 
     /**